refactor: update build config for @primer/styled-react to rolldown (#8043)
