+++ packages/styled-react/package.json
@@ -44,7 +44,6 @@
     "@primer/primitives": "10.x || 11.x",
     "@primer/react": "^38.26.0",
     "@rolldown/plugin-babel": "^0.2.3",
-    "@rollup/plugin-babel": "^6.1.0",
     "@storybook/react-vite": "^10.4.2",
     "@types/babel__core": "^7.20.5",
     "@types/react": "18.3.11",
@@ -59,8 +58,8 @@
     "react-compiler-runtime": "^1.0.0",
     "react-dom": "18.3.1",
     "rimraf": "^6.0.1",
-    "rollup": "4.59.0",
-    "rollup-plugin-typescript2": "^0.36.0",
+    "rolldown": "^1.1.2",
+    "rolldown-plugin-dts": "^0.26.0",
     "styled-components": "5.3.11",
     "typescript": "^6.0.3"
   },
