+++ packages/styled-react/script/build
@@ -1,10 +1,9 @@
 #!/bin/bash
 
-# Build the package
-npx rollup -c
+set -e
 
-# Generate type declarations
-npx tsc --project tsconfig.build.json
+# Build the package
+npx rolldown -c
 
 # Generate components.json
 ./script/generate-components-json
